added functional Books API to add a book

+++ app/src/main/java/com/cmput301/w19t06/theundesirablejackals/activities/MapsActivity.java
@@ -67,7 +67,7 @@ public class MapsActivity extends FragmentActivity implements OnMapReadyCallback
                         Double lt = geolocation.getLatitude();
                         Double ln = geolocation.getLongitude();
 
-                        coord = new LatLng(lt.intValue(), ln.intValue()); // set to some random location
+                        coord = new LatLng(lt, ln);
                     }
 //
                 } else {
